Adding a Makefile

+++ src/lib.c
@@ -10,7 +10,7 @@ void PrintIntro()
 void waitFor(unsigned int secs)
 {
     unsigned int retTime = time(0) + secs;
-    while (time(0) < retTime)
+    while ((unsigned int)time(0) < retTime)
     {
     }
 }
